+++ local-volume/provisioner/pkg/discovery/discovery_test.go
@@ -89,12 +89,14 @@ var testStorageClasses = []*storagev1.StorageClass{
 
 var scMapping = map[string]common.MountConfig{
 	"sc1": {
-		HostDir:  testHostDir + "/dir1",
-		MountDir: testMountDir + "/dir1",
+		HostDir:    testHostDir + "/dir1",
+		MountDir:   testMountDir + "/dir1",
+		VolumeMode: "Filesystem",
 	},
 	"sc2": {
-		HostDir:  testHostDir + "/dir2",
-		MountDir: testMountDir + "/dir2",
+		HostDir:    testHostDir + "/dir2",
+		MountDir:   testMountDir + "/dir2",
+		VolumeMode: "Block",
 	},
 }
 
@@ -121,7 +123,7 @@ func TestDiscoverVolumes_Basic(t *testing.T) {
 			{Name: "symlink2", Hash: 0x23645a36, VolumeType: util.FakeEntryBlock, Capacity: 100 * 1024 * 1024},
 		},
 		"dir2": {
-			{Name: "mount1", Hash: 0xa7aafa3c, VolumeType: util.FakeEntryFile},
+			{Name: "symlink1", Hash: 0x55d5adba, VolumeType: util.FakeEntryBlock},
 			{Name: "symlink2", Hash: 0x226458a3, VolumeType: util.FakeEntryBlock},
 		},
 	}
@@ -142,7 +144,7 @@ func TestDiscoverVolumes_BasicTwice(t *testing.T) {
 			{Name: "symlink2", Hash: 0x23645a36, VolumeType: util.FakeEntryBlock},
 		},
 		"dir2": {
-			{Name: "mount1", Hash: 0xa7aafa3c, VolumeType: util.FakeEntryFile},
+			{Name: "symlink1", Hash: 0x55d5adba, VolumeType: util.FakeEntryBlock},
 			{Name: "symlink2", Hash: 0x226458a3, VolumeType: util.FakeEntryBlock},
 		},
 	}
@@ -194,7 +196,7 @@ func TestDiscoverVolumes_NewVolumesLater(t *testing.T) {
 			{Name: "symlink2", Hash: 0x23645a36, VolumeType: util.FakeEntryBlock},
 		},
 		"dir2": {
-			{Name: "mount1", Hash: 0xa7aafa3c, VolumeType: util.FakeEntryFile},
+			{Name: "symlink1", Hash: 0x55d5adba, VolumeType: util.FakeEntryBlock},
 			{Name: "symlink2", Hash: 0x226458a3, VolumeType: util.FakeEntryBlock},
 		},
 	}
@@ -230,7 +232,7 @@ func TestDiscoverVolumes_CreatePVFails(t *testing.T) {
 			{Name: "mount2", Hash: 0x79412c38, VolumeType: util.FakeEntryFile},
 		},
 		"dir2": {
-			{Name: "mount1", Hash: 0xa7aafa3c, VolumeType: util.FakeEntryFile},
+			{Name: "mount1", Hash: 0x55d5adba, VolumeType: util.FakeEntryFile},
 			{Name: "mount2", Hash: 0x7c4130f1, VolumeType: util.FakeEntryFile},
 		},
 	}
@@ -272,7 +274,7 @@ func TestDiscoverVolumes_CleaningInProgress(t *testing.T) {
 			{Name: "symlink2", Hash: 0x23645a36, VolumeType: util.FakeEntryBlock, Capacity: 100 * 1024 * 1024},
 		},
 		"dir2": {
-			{Name: "mount1", Hash: 0xa7aafa3c, VolumeType: util.FakeEntryFile},
+			{Name: "symlink1", Hash: 0x55d5adba, VolumeType: util.FakeEntryBlock},
 			{Name: "symlink2", Hash: 0x226458a3, VolumeType: util.FakeEntryBlock},
 		},
 	}
@@ -283,7 +285,7 @@ func TestDiscoverVolumes_CleaningInProgress(t *testing.T) {
 			{Name: "mount1", Hash: 0xaaaafef5, VolumeType: util.FakeEntryFile, Capacity: 100 * 1024},
 		},
 		"dir2": {
-			{Name: "mount1", Hash: 0xa7aafa3c, VolumeType: util.FakeEntryFile},
+			{Name: "symlink1", Hash: 0x55d5adba, VolumeType: util.FakeEntryBlock},
 			{Name: "symlink2", Hash: 0x226458a3, VolumeType: util.FakeEntryBlock},
 		},
 	}
@@ -296,6 +298,38 @@ func TestDiscoverVolumes_CleaningInProgress(t *testing.T) {
 	// Mark dir1/mount2 PV as being cleaned. This one should not get created
 	pvName := getPVName(vols["dir1"][1])
 	test.cleanupTracker.ProcTable.MarkRunning(pvName)
+
+	d.DiscoverLocalVolumes()
+	verifyCreatedPVs(t, test)
+}
+
+func TestDiscoverVolumes_InvalidMode(t *testing.T) {
+	vols := map[string][]*util.FakeDirEntry{
+		"dir1": {
+			{Name: "mount1", Hash: 0xaaaafef5, VolumeType: util.FakeEntryFile, Capacity: 100 * 1024},
+			{Name: "symlink2", Hash: 0x23645a36, VolumeType: util.FakeEntryBlock, Capacity: 100 * 1024 * 1024},
+		},
+		"dir2": {
+			{Name: "mount1", Hash: 0xa7aafa3c, VolumeType: util.FakeEntryFile},
+			{Name: "symlink2", Hash: 0x226458a3, VolumeType: util.FakeEntryBlock},
+		},
+	}
+
+	// Don't expect dir2/mount1 to be created, due to invalid volume mode.
+	expectedVols := map[string][]*util.FakeDirEntry{
+		"dir1": {
+			{Name: "mount1", Hash: 0xaaaafef5, VolumeType: util.FakeEntryFile, Capacity: 100 * 1024},
+			{Name: "symlink2", Hash: 0x23645a36, VolumeType: util.FakeEntryBlock, Capacity: 100 * 1024 * 1024},
+		},
+		"dir2": {
+			{Name: "symlink2", Hash: 0x226458a3, VolumeType: util.FakeEntryBlock},
+		},
+	}
+	test := &testConfig{
+		dirLayout:       vols,
+		expectedVolumes: expectedVols,
+	}
+	d := testSetup(t, test, false)
 
 	d.DiscoverLocalVolumes()
 	verifyCreatedPVs(t, test)
@@ -358,15 +392,15 @@ func testSetup(t *testing.T, test *testConfig, useAlphaAPI bool) *Discoverer {
 	return d
 }
 
-func findSCName(t *testing.T, targetDir string, test *testConfig) string {
+func findSCNameAndVolumeMode(t *testing.T, targetDir string, test *testConfig) (string, string) {
 	for sc, config := range scMapping {
 		_, dir := filepath.Split(config.HostDir)
 		if dir == targetDir {
-			return sc
+			return sc, config.VolumeMode
 		}
 	}
 	t.Fatalf("Failed to find SC Name for directory %v", targetDir)
-	return ""
+	return "", ""
 }
 
 func verifyNodeAffinity(t *testing.T, pv *v1.PersistentVolume) {
@@ -461,12 +495,23 @@ func verifyCapacity(t *testing.T, createdPV *v1.PersistentVolume, expectedPV *te
 	}
 }
 
+func verifyVolumeMode(t *testing.T, createdPV *v1.PersistentVolume, expectedPV *testPVInfo) {
+	if createdPV.Spec.VolumeMode == nil {
+		t.Errorf("Unknown volume mode in created PV")
+	}
+
+	if *createdPV.Spec.VolumeMode != expectedPV.volumeMode {
+		t.Errorf("Expected mode %q, got %q", expectedPV.volumeMode, *createdPV.Spec.VolumeMode)
+	}
+}
+
 // testPVInfo contains all the fields we are intested in validating.
 type testPVInfo struct {
 	pvName       string
 	path         string
 	capacity     int64
 	storageClass string
+	volumeMode   v1.PersistentVolumeMode
 }
 
 func getPVName(entry *util.FakeDirEntry) string {
@@ -479,11 +524,13 @@ func verifyCreatedPVs(t *testing.T, test *testConfig) {
 		for _, file := range files {
 			pvName := getPVName(file)
 			path := filepath.Join(testHostDir, dir, file.Name)
+			sc, mode := findSCNameAndVolumeMode(t, dir, test)
 			expectedPVs[pvName] = &testPVInfo{
 				pvName:       pvName,
 				path:         path,
 				capacity:     file.Capacity,
-				storageClass: findSCName(t, dir, test),
+				storageClass: sc,
+				volumeMode:   v1.PersistentVolumeMode(mode),
 			}
 		}
 	}
@@ -516,6 +563,7 @@ func verifyCreatedPVs(t *testing.T, test *testConfig) {
 		verifyNodeAffinity(t, createdPV)
 		verifyPVLabels(t, createdPV)
 		verifyCapacity(t, createdPV, expectedPV)
+		verifyVolumeMode(t, createdPV, expectedPV)
 		// TODO: Verify volume type once that is supported in the API.
 	}
 }